LayoutTests/webrtc/video-stats.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <title>Testing basic video exchange from offerer to receiver</title>
        <script src="../resources/testharness.js"></script>
        <script src="../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script src ="routines.js"></script>
        <script>

function checkStatsReportIterator(report)
{
    assert_equals(Object.getOwnPropertyDescriptor(report.__proto__, Symbol.iterator).value, Object.getOwnPropertyDescriptor(report.__proto__, 'entries').value);
    assert_equals(Object.getOwnPropertyDescriptor(report.__proto__, Symbol.iterator).value.name, "entries");
    for (let pair of report)
        assert_equals(pair.length, 2);
}

function getInboundRTPStats(connection)
{
    return connection.getStats().then((report) => {
        checkStatsReportIterator(report);
        var stats;
        report.forEach((statItem) => {
            if (statItem.type === "inbound-rtp") {
                stats = statItem;
            }
        });
        return stats;
    });
}

function getOutboundRTPStats(connection)
{
    return connection.getStats().then((report) => {
        checkStatsReportIterator(report);
        var stats;
        report.forEach((statItem) => {
            if (statItem.type === "outbound-rtp") {
                stats = statItem;
            }
        });
        return stats;
    });
}

var firstConnection, secondConnection;
promise_test((test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);

    var localStream, remoteStream;
    return navigator.mediaDevices.getUserMedia({ video: true}).then((stream) => {
        localStream = stream;
        return new Promise((resolve, reject) => {
            createConnections((connection) => {
                firstConnection = connection;
                firstConnection.addTrack(stream.getVideoTracks()[0], stream);
            }, (connection) => {
                secondConnection = connection;
                secondConnection.ontrack = (trackEvent) => {
                    remoteStream = trackEvent.streams[0];
                    resolve();
                };
            });
            setTimeout(() => reject("Test timed out"), 5000);
        });
    }).then(() => {
        return getOutboundRTPStats(firstConnection);
    }).then((stats) => {
        assert_true(!!stats, "outbound-rtp stats should not be null");
        assert_true(Number.isInteger(stats.framesEncoded), "framesEncoded should be an integer");
        assert_true(Number.isInteger(stats.qpSum), "outbound qpSum should be an integer");
        statsFirstConnection = stats;
        return getInboundRTPStats(secondConnection);
    }).then((stats) => {
        assert_true(!!stats, "inbound-rtp stats should not be null");
        assert_true(Number.isInteger(stats.framesDecoded), "framesDecoded should be an integer");
        assert_true(Number.isInteger(stats.qpSum), "inbound qpSum should be an integer");
        statsSecondConnection = stats;
        return waitFor(300);
    }).then(() => {
        return getOutboundRTPStats(firstConnection);
    }).then((stats) => {
        assert_true(stats.timestamp > statsFirstConnection.timestamp, "Timestamp for first connection should have increased");
        assert_true(stats.framesEncoded > statsFirstConnection.framesEncoded, "Number of encoded frames should have increased");

        return getInboundRTPStats(secondConnection);
    }).then((stats) => {
        assert_true(stats.timestamp > statsSecondConnection.timestamp, "Timestamp for second connection should have increased");
        assert_true(stats.framesDecoded > statsSecondConnection.framesDecoded, "Number of decoded frames should have increased");
    });
}, "Basic video stats");
        </script>
    </body>
</html>
